Code refactoring, DateUtil now consists of all Date and time related functions
