+++ Hydrate/lint.xml
@@ -1,3 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<lint>
-</lint>
+<lint></lint>
