+++ Hydrate/src/main/AndroidManifest.xml
@@ -199,13 +199,13 @@
             android:label="@string/initial_setup" />
 
         <!--<receiver android:name=".widgets.ProgressWidget">-->
-            <!--<intent-filter>-->
-                <!--<action android:name="android.appwidget.action.APPWIDGET_UPDATE" />-->
-            <!--</intent-filter>-->
+        <!--<intent-filter>-->
+        <!--<action android:name="android.appwidget.action.APPWIDGET_UPDATE" />-->
+        <!--</intent-filter>-->
 
-            <!--<meta-data-->
-                <!--android:name="android.appwidget.provider"-->
-                <!--android:resource="@xml/progress_widget_info" />-->
+        <!--<meta-data-->
+        <!--android:name="android.appwidget.provider"-->
+        <!--android:resource="@xml/progress_widget_info" />-->
         <!--</receiver>-->
     </application>
