+++ Hydrate/src/main/java/com/underdog/hydrate/preference/TargetQuantityEditTextPreference.java
@@ -15,7 +15,7 @@ import com.underdog.hydrate.constants.Constants;
 import com.underdog.hydrate.database.HydrateContentProvider;
 import com.underdog.hydrate.database.HydrateDatabase;
 import com.underdog.hydrate.fragments.DayScheduleDialog;
-import com.underdog.hydrate.util.Utility;
+import com.underdog.hydrate.util.DateUtil;
 
 public class TargetQuantityEditTextPreference extends Preference {
     SharedPreferences defaultPreferences = PreferenceManager
@@ -52,7 +52,7 @@ public class TargetQuantityEditTextPreference extends Preference {
         String value;
         targetCursor = getContext().getContentResolver().query(HydrateContentProvider.CONTENT_URI_HYDRATE_DAILY_SCHEDULE,
                 new String[]{HydrateDatabase.COLUMN_TARGET_QUANTITY}, HydrateDatabase.DAY + "=?",
-                new String[]{Utility.getInstance().getToday() + ""}, null);
+                new String[]{DateUtil.getInstance().getToday() + ""}, null);
         targetCursor.moveToFirst();
         value = targetCursor.getDouble(0) + "";
         setSummary(value);
