+++ Hydrate/src/main/res/layout/activity_setup.xml
@@ -260,16 +260,16 @@
         app:layout_constraintVertical_bias="1.0">
 
         <!--<Button-->
-            <!--android:id="@+id/setupSkip"-->
-            <!--android:layout_width="100dp"-->
-            <!--android:layout_height="wrap_content"-->
-            <!--android:layout_marginEnd="8dp"-->
-            <!--android:layout_weight="0.3"-->
-            <!--android:background="@color/danger"-->
-            <!--android:text="@string/skip"-->
-            <!--android:textColor="@color/white"-->
-            <!--app:layout_constraintLeft_toLeftOf="parent"-->
-            <!--tools:layout_editor_absoluteY="539dp" />-->
+        <!--android:id="@+id/setupSkip"-->
+        <!--android:layout_width="100dp"-->
+        <!--android:layout_height="wrap_content"-->
+        <!--android:layout_marginEnd="8dp"-->
+        <!--android:layout_weight="0.3"-->
+        <!--android:background="@color/danger"-->
+        <!--android:text="@string/skip"-->
+        <!--android:textColor="@color/white"-->
+        <!--app:layout_constraintLeft_toLeftOf="parent"-->
+        <!--tools:layout_editor_absoluteY="539dp" />-->
 
         <Button
             android:id="@+id/setupSubmit"
