+++ Hydrate/src/main/res/layout/fragment_backup.xml
@@ -6,6 +6,6 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".BackupActivity$PlaceholderFragment" >
+    tools:context=".BackupActivity$PlaceholderFragment">
 
 </LinearLayout>
