+++ Hydrate/src/main/res/layout/fragment_dnd.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.underdog.hydrate.DndCardViewFragment" >
+    tools:context="com.underdog.hydrate.DndCardViewFragment">
 
     <Button
         android:layout_width="match_parent"
