+++ Hydrate/src/main/res/layout/fragment_edit_cups.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
-    tools:context="com.underdog.hydrate.EditCupsActivity$PlaceholderFragment" >
+    tools:context="com.underdog.hydrate.EditCupsActivity$PlaceholderFragment">
 
     <LinearLayout
         android:id="@+id/adLayoutEdit"
@@ -12,8 +12,7 @@
         android:layout_height="wrap_content"
         android:layout_alignParentTop="true"
         android:layout_centerHorizontal="true"
-        android:orientation="vertical" >
-    </LinearLayout>
+        android:orientation="vertical"></LinearLayout>
 
     <ListView
         android:id="@+id/editCupsListView"
@@ -21,7 +20,6 @@
         android:layout_height="match_parent"
         android:layout_below="@id/adLayoutEdit"
         android:focusable="false"
-        android:scrollbars="none" >
-    </ListView>
+        android:scrollbars="none"></ListView>
 
 </RelativeLayout>
