+++ Hydrate/src/main/res/layout/fragment_summary.xml
@@ -12,8 +12,7 @@
         android:layout_height="wrap_content"
         android:layout_alignParentTop="true"
         android:layout_centerHorizontal="true"
-        android:orientation="vertical" >
-    </LinearLayout>
+        android:orientation="vertical"></LinearLayout>
 
     <TextView
         android:id="@+id/averageConsumption"
@@ -39,7 +38,6 @@
         android:layout_alignParentStart="true"
         android:layout_below="@id/averageConsumption"
         android:footerDividersEnabled="false"
-        android:headerDividersEnabled="false">
-    </ListView>
+        android:headerDividersEnabled="false"></ListView>
 
 </RelativeLayout>
