+++ Hydrate/src/main/res/layout/instant_mute_dialog.xml
@@ -2,7 +2,7 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/grey" >
+    android:background="@color/grey">
 
     <TextView
         android:id="@+id/instant_mute_heading"
@@ -21,4 +21,5 @@
         android:layout_height="wrap_content"
         android:layout_below="@+id/instant_mute_heading"
         android:layout_centerHorizontal="true" />
-</RelativeLayout>
+
+</RelativeLayout>
