+++ Hydrate/src/main/res/layout/progress_layout.xml
@@ -5,7 +5,7 @@
     android:gravity="center"
     android:orientation="horizontal"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingTop="@dimen/activity_vertical_margin" >
+    android:paddingTop="@dimen/activity_vertical_margin">
 
     <ProgressBar
         android:id="@+id/progress"
