+++ Hydrate/src/main/res/layout/today_log_layout.xml
@@ -3,7 +3,7 @@
     android:id="@+id/todayLogLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="horizontal" >
+    android:orientation="horizontal">
 
     <TextView
         android:id="@+id/_id"
