+++ Hydrate/src/main/res/menu/main.xml
@@ -2,11 +2,11 @@
     tools:context="com.underdog.hydrate.MainActivity">
 
     <!--<item-->
-        <!--android:id="@+id/menu_calendar"-->
-        <!--android:icon="@drawable/ic_menu_calendar"-->
-        <!--android:orderInCategory="100"-->
-        <!--android:title="@string/instant_mute"-->
-        <!--app:showAsAction="always" />-->
+    <!--android:id="@+id/menu_calendar"-->
+    <!--android:icon="@drawable/ic_menu_calendar"-->
+    <!--android:orderInCategory="100"-->
+    <!--android:title="@string/instant_mute"-->
+    <!--app:showAsAction="always" />-->
     <!--<item-->
     <!--android:id="@+id/instant_mute"-->
     <!--android:icon="@drawable/ic_action_bell"-->
