+++ Hydrate/src/main/res/values/keys.xml
@@ -29,6 +29,4 @@
     <!-- Key strings.Not to be used in views for display -->
 
 
-
-
 </resources>
