+++ Hydrate/src/main/res/xml/progress_widget_info.xml
@@ -6,4 +6,4 @@
     android:minWidth="250dp"
     android:previewImage="@drawable/example_appwidget_preview"
     android:updatePeriodMillis="86400000"
-    android:widgetCategory="home_screen"/>
+    android:widgetCategory="home_screen" />
